+++ src/wrapper.rs
@@ -20,6 +20,8 @@
 ///
 /// Trait defines convenient methods for accessing inner data, construct
 /// and deconstruct newtype. It also serves as a marker trait for newtypes.
+///
+/// The trait works well with `#[derive(Wrapper)]` from `amplify_derive` crate
 pub trait Wrapper {
     /// Inner type wrapped by the current newtype
     type Inner: Clone;
@@ -30,6 +32,10 @@ pub trait Wrapper {
     /// Returns reference to the inner representation for the wrapper type
     fn as_inner(&self) -> &Self::Inner;
 
+    /// Returns a mutable reference to the inner representation for the wrapper
+    /// type
+    fn as_inner_mut(&mut self) -> &mut Self::Inner;
+
     /// Clones inner data of the wrapped type and return them
     #[inline]
     fn to_inner(&self) -> Self::Inner {
@@ -38,110 +44,4 @@ pub trait Wrapper {
 
     /// Unwraps the wrapper returning the inner type
     fn into_inner(self) -> Self::Inner;
-}
-
-// TODO: Add generic support to the wrapper
-// TODO: Convert to derive macro
-/// Macro simplifying creation of new wrapped types. It automatically implements
-/// [`Wrapper`] trait, adds default implementation for the following traits:
-/// * [`AsRef`]
-/// * [`AsMut`]
-/// * [`Borrow`]
-/// * [`BorrowMut`]
-/// * [`Deref`]
-/// * [`DerefMut`]
-/// * [`From`]`<Wrapper>`
-/// * [`From`]`<Inner>`
-///
-/// Macro allows to add custom derives to the newtype using `derive` argument
-#[macro_export]
-macro_rules! wrapper {
-    ($name:ident, $from:ty, $docs:meta, derive=[$( $derive:ident ),+]) => {
-        #[$docs]
-        #[derive(Clone, Debug)]
-        $( #[derive($derive)] )+
-        pub struct $name($from);
-
-        impl $crate::Wrapper for $name {
-            type Inner = $from;
-
-            #[inline]
-            fn from_inner(inner: $from) -> Self {
-                Self(inner)
-            }
-
-            #[inline]
-            fn as_inner(&self) -> &$from {
-                &self.0
-            }
-
-            #[inline]
-            fn into_inner(self) -> $from {
-                self.0
-            }
-        }
-
-        impl ::core::convert::AsRef<$from> for $name {
-            #[inline]
-            fn as_ref(&self) -> &$from {
-                &self.0
-            }
-        }
-
-        impl ::core::convert::AsMut<$from> for $name {
-            #[inline]
-            fn as_mut(&mut self) -> &mut $from {
-                &mut self.0
-            }
-        }
-
-        impl ::core::borrow::Borrow<$from> for $name {
-            #[inline]
-            fn borrow(&self) -> &$from {
-                &self.0
-            }
-        }
-
-        impl ::core::borrow::BorrowMut<$from> for $name {
-            #[inline]
-            fn borrow_mut(&mut self) -> &mut $from {
-                &mut self.0
-            }
-        }
-
-        impl ::core::ops::Deref for $name {
-            type Target = $from;
-            #[inline]
-            fn deref(&self) -> &Self::Target {
-                &self.0
-            }
-        }
-
-        impl ::core::ops::DerefMut for $name {
-            #[inline]
-            fn deref_mut(&mut self) -> &mut Self::Target {
-                &mut self.0
-            }
-        }
-
-        impl ::core::convert::From<$from> for $name
-        where
-            Self: ::core::clone::Clone,
-        {
-            #[inline]
-            fn from(x: $from) -> Self {
-                Self(x)
-            }
-        }
-
-        impl ::core::convert::From<&$from> for $name
-        where
-            Self: ::core::clone::Clone,
-        {
-            #[inline]
-            fn from(x: &$from) -> Self {
-                Self(x.clone())
-            }
-        }
-    };
 }